test(tests/conftest): move tmp_git_project and tmp_commitizen_project into conftest
they're commonly used fixture among all tests
